Added exp dao

+++ app/src/main/java/talkapp/org/talkappmobile/activity/listener/OnMainActivityListener.java
@@ -3,5 +3,5 @@ package talkapp.org.talkappmobile.activity.listener;
 public interface OnMainActivityListener {
     void onAppVersionInitialized(String packageName);
 
-    void onYourExpInitialized(int exp);
+    void onYourExpInitialized(double exp);
 }
