+++ app/src/main/java/talkapp/org/talkappmobile/activity/presenter/MainActivityPresenter.java
@@ -31,7 +31,7 @@ public class MainActivityPresenter implements OnMainActivityListener {
     }
 
     @Override
-    public void onYourExpInitialized(int exp) {
+    public void onYourExpInitialized(double exp) {
         view.setYourExp("EXP " + exp);
     }
 }
